+++ crates/pdk-test-utils/src/lib.rs
@@ -1,8 +1,10 @@
 #![allow(clippy::disallowed_types)]
 
+mod extension_wrapper;
 mod sandbox;
-mod wrappers;
+mod toolchain_wrapper;
 
+pub use extension_wrapper::*;
 pub use moon_pdk_api::*;
 pub use sandbox::*;
-pub use wrappers::*;
+pub use toolchain_wrapper::*;